+++ utils/State.h
@@ -13,16 +13,16 @@
 struct State {
     std::string expr;
     int maxIters;
-    float metricArg;
-    std::complex<float> p{0};
+    double metricArg;
+    std::complex<double> p{0};
     Viewport viewport;
     bool colorCutoffEnabled;
-    float colorCutoff;
+    double colorCutoff;
     int width, height;
     ModeInfo mode; //Only the DistanceMetric is serialized
-    std::optional<std::complex<float>> pathStart = {};
-    std::optional<std::complex<float>> lineTransStart = {};
-    std::optional<std::complex<float>> lineTransEnd = {};
+    std::optional<std::complex<double>> pathStart = {};
+    std::optional<std::complex<double>> lineTransStart = {};
+    std::optional<std::complex<double>> lineTransEnd = {};
     int lineTransIteration = 0;
     bool lineTransEnabled = false; // Never serialised, inferred from lineTransEnd during deserialization
     bool forceDisableIncrementalLineTracing = false;
@@ -59,8 +59,8 @@ struct State {
             ar & lineTransStart;
             ar & lineTransEnd;
         } else {
-            ar & std::optional<std::complex<float>>();
-            ar & std::optional<std::complex<float>>();
+            ar & std::optional<std::complex<double>>();
+            ar & std::optional<std::complex<double>>();
         }
         ar & forceDisableIncrementalLineTracing;
     }
